Move Python 3 syntax tests into main tests
Previously these were separated to support Python 2.
Now we can include them in normal tests
